refactor: fix typos and imports
